Move integration test files into new structure
