+++ extension/src/test/suite/cli/dvc/runner.test.ts
@@ -2,15 +2,15 @@ import { afterEach, beforeEach, describe, it, suite } from 'mocha'
 import { expect } from 'chai'
 import { restore, spy, stub } from 'sinon'
 import { window, Event, EventEmitter } from 'vscode'
-import { Disposable, Disposer } from '../../../extension'
-import { Config } from '../../../config'
-import { DvcRunner } from '../../../cli/dvc/runner'
-import { CliResult, CliStarted } from '../../../cli'
-import * as Telemetry from '../../../telemetry'
-import { EventName } from '../../../telemetry/constants'
-import { WEBVIEW_TEST_TIMEOUT } from '../timeouts'
+import { Disposable, Disposer } from '../../../../extension'
+import { Config } from '../../../../config'
+import { DvcRunner } from '../../../../cli/dvc/runner'
+import { CliResult, CliStarted } from '../../../../cli'
+import * as Telemetry from '../../../../telemetry'
+import { EventName } from '../../../../telemetry/constants'
+import { WEBVIEW_TEST_TIMEOUT } from '../../timeouts'
 
-suite('CLI Runner Test Suite', () => {
+suite('DVC Runner Test Suite', () => {
   const disposable = Disposable.fn()
 
   beforeEach(() => {
@@ -21,7 +21,7 @@ suite('CLI Runner Test Suite', () => {
     disposable.dispose()
   })
 
-  describe('dvcRunner', () => {
+  describe('DvcRunner', () => {
     it('should only be able to run a single command at a time', async () => {
       const dvcRunner = disposable.track(new DvcRunner({} as Config, 'sleep'))
